_common/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "helper"
version = "0.1.0-dev0"
description = "Common helper functions for AWS tools local development"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Kay Hau", email = "virtualda@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
    "boto3==1.35.19",
    "click==8.1.7",
    "docker==7.1.0",
    "InquirerPy==0.3.4",
    "prompt-toolkit>=3.0.13",
    "pyyaml==6.0.2",
]

[project.urls]
Homepage = "https://github.com/kyhau/aws-tools/"
Repository = "https://github.com/kyhau/aws-tools/"

[project.scripts]
helper = "helper:main"
dockerc = "helper.docker:find_non_running_containers"
dockeri = "helper.docker:find_dangling_images"

# Poetry-specific configuration
[tool.poetry]
packages = [{include = "helper"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
boto3 = "1.35.19"
click = "8.1.7"
docker = "7.1.0"
InquirerPy = "0.3.4"
prompt-toolkit = ">=3.0.13"
pyyaml = "6.0.2"

[tool.poetry.group.test.dependencies]
pytest = "~8.4"
pytest-cov = "~7.0"
pytest-mock = "~3.15"

[tool.poetry.group.dev.dependencies]
black = "~25.9"
flake8 = "~7.3"
yamllint = "~1.35"

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = [".tox", "env", ".eggs", "dist", "build"]
xfail_strict = true
addopts = "--strict-markers"

# Coverage configuration
[tool.coverage.run]
source = ["helper"]
omit = [
    "setup.py",
    "*__init__.py",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
